+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.5] - 2020-01-09
+### Added
+ - strings::significant_terms function that extract significant terms from a data subset like elasticsearch 
+ significant_text aggregation.
+### Changed
+- confusion matrix image in read me converted to a github link.
+- strings::append_tags_to_frame added parameters max_features, lowercase and tokenizer.
+### Fixed
+- minor changes
+
 ## [1.4.1] - 2019-12-26
 ### Added